requirements.txt
openai
chromadb
python-dotenv
fastapi
uvicorn